Fix a bug where when using /cprivate manually the null would not be removed from the cache:
There was already checks to remove the nulled block but to get the Protection object after creating
a protection, loadProtection(world, x, y, z) is used. This respects the null cache, so it always
returned null when using /cprivate, which is obviously bad. We instead remove the key from the
cache before using loadProtection () when creating a new protection so that this cannot happen.