The rest of the cache synchronization issues should* now be fixed.
The problem lied in protected blocks that DON'T contain the protection being
assumed to have a null protection. This caused a disconnect between what
blocks should be protected and which really are. So we force the cache to
be updated instead of only when it is first looked up. No large performance
impact.