Experimentation with gettext -- backend module will generate .keys and .po files for both api/ and backend/ (although api has minimal usages of i18n, mainly just errors). No real implementation internally just yet ! :)
The current way LWC v4 does translations is a tad messy and difficult to update. As well, messages in the source are unreadable. gettext is excellent and definitely a trusted and solid way to do translations and as such I should not be reinventing the wheel with something that has many great tools out there for it to aid in translating.