Add a db pool in a slightly elegant way without having to modify the existing code base TOO much!:
Database pools required you to get a Connection object and then close it afterwards. LWC has 2000+ lines of database code. Modifying every instance of this is extremely (!!) time consuming and very inconvenient.

Instead of doing this we exploit the fact that in LWC's usecase, once a PreparedStatement or ResultSet is closed, everything above it is closed, e.g a close chain: ResultSet -> PreparedStatement -> Connection. This is done automatically by the new backing AutoClosingPreparedStatement and AutoClosingResultSet.

This however did require manually modifying anything that used raw Statement objects.

This means that this build truly is a "dev" build not an unreleased stable build. Care should be taken with this build.