Highly experimental entity protection & slightly improved magnets. NOTE THE FOLLOWING:
THIS HAS NOT BEEN FULLY TESTED OR EXPLORED.

This extends LWC protections to the following major entities: Storage Minecarts, Portraits, and Item Frames. LWC has been originally designed to work with clean Block & Material objects. Making it work with entities is interesting so I _expect bugs_!

* All wall entities should be protectable: portraits, item frames
* => To interact with them you need to try to destroy them.
* => You yourself cannot destroy a protected entity. You need to /cremove it
* => Protected entities will NEVER BE DESTROYED unless /cremove'd at the moment. That means if the wall backing them is removed, the entity stays! And if water fills your beautiful palace, the entities again stay.
* STORAGE MINECARTS are protectable
* => Furthermore, /cmagnet WILL work on storage minecarts
* => When interacting with the cart, try to open the inventory (right click)
* /lwc admin cleanup has been tested and properly checks for entities but has not been fully examined
* There *may* be entities that I don't know of that can be protected. I've tried protecting MOBs and it didn't work so you don't need to try that! (although LWC could probably protect mobs really easily now, but that's a different story)

Again, bugs are expected at this point so PLEASE DO NOT RUN ON HUGE 300+ PLAYER PRODUCTION SERVERS! As much as I would love that, you'll probably hit a brick wall very quickly.

This addresses #392 for the most part.

There is s till more left to do for a full implementation however this is a good start.