Finally only download libraries when actually needed. It is done by checking for a specific class from that library. If it's not available then the library will be loaded (and downloaded if not already downloaded).
This is all done via a new resources.yml. A base path for the files to be downloaded from can also be set. The resourcelibs.txt file now contains the absolute minimum required files for LWC to startup (YAML & gettext). Just database libraries are in resources.yml at the moment.

As well, Minecraft (which Forge uses for mods after the bootstrap phase) has its own class loader that caches classes that were previously not loaded when trying to load a class. Because I check the class _and then_ load it this means the class can never actually be loaded. To get around this I've reflected into the class and removed the class I checked. It is just simply a Set<String>