Prepare for adaptive cache testing with a known-nulls cache. This existed probably over a year ago as well but was removed. Basically:
It is just a cache of blocks that are known to not have any protections on them. If all protections are not in the cache then an otherwise cache miss would force a DB lookup. This prevents that DB lookup after the first time it's done (and nothing is found)