#811 Initial UUID work. This is NOT backwards compatible with MC 1.6
Offline mode: Names shouldn't be converted to the UUID of the user with the same username (online mode) as long as the player already logged into the server at least once.
Offline mode might still have a couple bugs as a result.

* When a protection is loaded: Check if it needs converting; if it does, convert it
* Use UUIDs when creating protections and modifying rights
* /cadmin find now searches by UUID
* Change output from varous commands to attempt to resolve a player's name instead of using a UUID

This does not yet slowly crawl the database to convert everything.