New string of deprecations in the first steps towards improving how protected blocks are stored.
At the moment (and in LWCv4) only 1 block in the protection is stored. When a blocked is clicked in the world, an expensive search is performed to find protectable blocks nearby that the block could be part of and then queries the database for each possible protectable block. This performance can be improved with caching.

The aim is to instead store all protected blocks in the database. This will result in the following (ignoring caching):

* Seeing if a block is protected only requires 1 database lookup on the block
* No blocks other than the queried block is looked at in the world (i.e. no adjacent blocks are queried)
* The somewhat expensive world search is only done when a protection is created