Reintroduce functionality from the old any interact that told the player when they clicked on the wrong block.
i.e., if they are told to "Click on a protection" but they click on a block that is not protected, they will be told that it is not protected and to click on a block that is.

The functionality is automatic when using either player.onNextProtectionInteract or player.onNextBlockInteract.