Forge: Load mappings file for the current MC version from the update server.
One major benefit of this is that if LWC version X for MC 1.7.10 is placed on a server with MC 1.8, it's essentially guaranteed to not work (no correct mappings). If version X can pull the mappings for the version it's on (MC 1.8), it has a much better chance of actually working.